libdeflate update

+++ 2.0/libdeflate/lib/deflate_compress.c
@@ -533,7 +533,7 @@ deflate_flush_bits(struct deflate_output_bitstream *os)
 		/* Flush a whole word (branchlessly).  */
 		put_unaligned_leword(os->bitbuf, os->next);
 		os->bitbuf >>= os->bitcount & ~7;
-		os->next += MIN((u32)(os->end - os->next), os->bitcount >> 3);
+		os->next += MIN(os->end - os->next, os->bitcount >> 3);
 		os->bitcount &= 7;
 	} else {
 		/* Flush a byte at a time.  */
@@ -1664,7 +1664,7 @@ deflate_write_uncompressed_block(struct deflate_output_bitstream *os,
 				   DEFLATE_BLOCKTYPE_UNCOMPRESSED);
 	deflate_align_bitstream(os);
 
-	if (4 + (u32)len >= (u32)(os->end - os->next)) {
+	if (4 + (u32)len >= os->end - os->next) {
 		os->next = os->end;
 		return;
 	}
@@ -1996,7 +1996,7 @@ deflate_compress_greedy(struct libdeflate_compressor * restrict c,
 
 			/* Decrease the maximum and nice match lengths if we're
 			 * approaching the end of the input buffer.  */
-			if (unlikely(max_len > (u32)(in_end - in_next))) {
+			if (unlikely(max_len > in_end - in_next)) {
 				max_len = in_end - in_next;
 				nice_len = MIN(nice_len, max_len);
 			}
@@ -2527,7 +2527,7 @@ deflate_compress_near_optimal(struct libdeflate_compressor * restrict c,
 
 			/* Decrease the maximum and nice match lengths if we're
 			 * approaching the end of the input buffer.  */
-			if (unlikely(max_len > (u32)(in_end - in_next))) {
+			if (unlikely(max_len > in_end - in_next)) {
 				max_len = in_end - in_next;
 				nice_len = MIN(nice_len, max_len);
 			}
@@ -2599,7 +2599,7 @@ deflate_compress_near_optimal(struct libdeflate_compressor * restrict c,
 						in_next_slide = in_next + MIN((u32)(in_end - in_next),
 									      MATCHFINDER_WINDOW_SIZE);
 					}
-					if (unlikely(max_len > (u32)(in_end - in_next))) {
+					if (unlikely(max_len > in_end - in_next)) {
 						max_len = in_end - in_next;
 						nice_len = MIN(nice_len, max_len);
 					}
